put index functions in separate file
